feat: Add MongoDB dependency and update Prisma client initialization
This commit adds the MongoDB dependency to the project by including it in the package.json file. Additionally, it updates the initialization of the Prisma client in the prismadb.ts file to use a singleton pattern. This change ensures that only one instance of the Prisma client is created and used throughout the application.

Recent repository commits:
- Update ProposalBtn to fetch user data, Add validation for user data in ProposalBtn
- removed unnecessary components from navbar
- Add executive summary field to CompanyInfoForm and implement TeamMembersForm
- show proposals and update them
- standardised the file locations
- added vercel prisma script
- Update dependencies and fixed building bugs
- Update dependencies and add Quill CSS for text editor
- added functioning text editor
- added proposal templates and created proposal routes